move the logic to check for db connection outside the database package
