+++ src/cmd/api/main.go
@@ -1,14 +1,25 @@
 package main
 
 import (
+	"context"
+
 	"github.com/Real-Dev-Squad/reciprocal-backend/src/internal/config"
+	"github.com/Real-Dev-Squad/reciprocal-backend/src/internal/database"
 	"github.com/Real-Dev-Squad/reciprocal-backend/src/internal/logger"
 	"github.com/Real-Dev-Squad/reciprocal-backend/src/internal/server"
 )
 
 func main() {
-	server := server.NewServer()
+	// Create a context that we can cancel
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
 
+	// create a database connection pool
+	dbPool := database.New(ctx)
+	defer dbPool.Close()
+
+	// Create a new server
+	server := server.NewServer(dbPool)
 	logger.Info("Server running on port", config.Port)
 	if err := server.ListenAndServe(); err != nil {
 		panic(err)
