+++ src/internal/server/server.go
@@ -6,23 +6,19 @@ import (
 	"time"
 
 	"github.com/Real-Dev-Squad/reciprocal-backend/src/internal/config"
-	"github.com/Real-Dev-Squad/reciprocal-backend/src/internal/database"
+	"github.com/jackc/pgx/v5/pgxpool"
 )
 
 type Server struct {
-	port int
-	db   database.Service
+	db *pgxpool.Pool
 }
 
-func NewServer() *http.Server {
-	NewServer := &Server{
-		port: config.Port,
-		db:   database.New(),
-	}
+func NewServer(db *pgxpool.Pool) *http.Server {
+	NewServer := &Server{db}
 
 	// Declare Server config
 	server := &http.Server{
-		Addr:         fmt.Sprintf(":%d", NewServer.port),
+		Addr:         fmt.Sprintf(":%d", config.Port),
 		Handler:      NewServer.RegisterRoutes(),
 		IdleTimeout:  time.Minute,
 		ReadTimeout:  10 * time.Second,
